+++ lib/url_image.dart
@@ -3,4 +3,5 @@ export 'src/config.dart';
 export 'src/downloader.dart';
 export 'src/file_storage.dart';
 export 'src/url_image.dart';
-export 'src/vector_graphics.dart';
+export 'src/vector/cache.dart';
+export 'src/vector/graphics.dart';