+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.2.0
+
+* The image `fit` is nullable now and its value is `null` by default.
+* Fixed the size of vector images with `BoxFit.none`.
+* Added the `animationFit` property. It can be helpful when placing the image in a [Column] or a [Row] with [CrossAxisAlignment.stretch] to properly align animated images with different sizes.
+* The `ink` property is now `false` by default.
+* Reduced image blinking on hot initialization (from cache).
+* Rasterized vector images are now cached in the separate [VectorImageCache] instead of the common `PaintingBinding.instance.imageCache` to eliminate image eviction due to cache overflow and to reduce blinking of vector images during size animation in the [Hero] widget.
+
 ## 1.1.0
 
 * Image layout and alignment have been fixed. [UrlImage] now has its own size based on the source image size when placed in an unconstrained area like [Column] and [Row].